// src/types/jobListType.ts
export type jobsListType = {
    id?:number
  companyName: string;
  website: string;
  title: string;
  jobType: string;
  benefits: string;
  state: string;
  location: string;
  experience: string;
  salary: string;
  tags: string;
  description: string;
  noOfOpenings: number;
educationQualification: string[];
};